Add support for pairing
Should work by using atvremote:

    atvremote pair

And then add the remote on the Apple TV. Default pin is 1234 (which
must be correctly entered) and name is "pyatv remote". These settings
can be changed with parameters to atvremote. Currently a pairing
timer is set to 60 seconds, so the process must be finished within
that time.