Beginning of support for play_url
A polling loop is used to keep the connection alive, which is not a nice
solution. But it will have to do until I find a better way. This can be
tested with atvremote:

atvremote --debug -a play_url=http://www.sample-videos.com/video/mp4/720/big_buck_bunny_720p_1mb.mp4,0

The last 0 is start time (interval 0..1, float, for some reason).

Tests are missing in this commit.