Extract AirPlay code to separate interface
AirPlay will play a bigger part in the future, e.g. with device
authentication. Adding new features in a good way with current API is
pretty tough, so this change aims to make that easier by extracting
AirPlay to a separate public interface (similar to remote_control or
playing).

THIS IS A BREAKING CHANGE! The play_url has moved from remote_control
to airplay and all users must update their code.