Beginning of refactor for MediaRemote protocol
This is the start of moving the DAAP protocol logic into its own place
so that MediaRemote protocol can be implemented alongside it. More
changes must be made, for instance, a new fake Apple TV must be created
so that functional tests can be shared between the two protocol
implementations. Also, the AirPlay parts in the test suite should be
extracted to its own FakeAirPlayDevice (or similiar) to make re-use of
tests easier.