Initial MRP implementation
This is the initial implementation of the MRP protocol. It's pretty
rough and not everything works, but it's a start. No tests and no
up-to-date documentation yet.

To try it out, just use:

   atvremote --debug --protocol mrp -a pair

Add --address in case you want to target a specific device (but you
must still use -a to discover the correct port).