Fixes broken device pairing
Pairing with DMAP and MRP is a bit different since MRP requires more
details about the device when pairing (both an address and port). This
kinda got lost in the implementation so pairing stopped working. With
this commit pairing should work again and the implementation should
also be simpler (no need to make it more general than it has to be).