Add proxy MRP script
This script will mimic an MRP device that you can pair with, using for instance
the iOS app. It will then relay messages to a real device, doing re-encryption
on the fly so that messages can be intercepted and read in plain text. Usage
instructions and caveats are placed as comments in the script.