Deprecate python 3.4 and support 3.7
aiohttp does not support python 3.4 in newer versions which results in
a lot of problems when dealing with that version (async/await are not
supported in python 3.4 which aiohttp uses). So this commit
unfortunately deprecates support for python 3.4 in order to make it
possible to maintain.

No syntax migration has been made, i.e. using async/await syntax, but
that switch has already been made on master.

This commit also fixes some asyncio errors mentioned in #150 as well
as #151.