Use iPhone as device type when pairing
During pairing, pyatv reports back to the Apple TV that it is an iPod.
This is mostly becuase all other libraries out there does that. But for
some reason tvOS will now remove all paired devices with the ipod type
(probably other as well - no tested). But when changed to iPhone
instead, like the iTunes Remote app sends, it sticks like it used to. So
chaning this to make pairing useful again.