Fix push updater in MRP
When splitting pairing and connecting, the flow when setting up push
updates changed as well (implicitly). Simply because a connection is
established with the device before it's possible to set the push update
listener, so potentially no callback would be received. Now, a call to
the listener is made when setting a new listener.