Fix flaky home button test in MRP
Because home and home_hold mapped to the same usage_page and usage code,
it would be seemingly "random" which button was detected in the fake
MRP device. Since python 3.5 maintains order, home would be found first
and tests succeed. The order is not necessarily maintained in earlier
versions of python, so test would fail sometimes.