Some updates to the MRP proxy
A new proxy instance is now created for every new client, so it is
possible to connect to the proxy multiple times. Pairing and enabling
of encryption also works as expected now and since the encryption seed
is hardcoded, there is no need to re-pair anymore. Pair once and it
will continue to work even when restarting the script. But no security
whatsoever.