Improve device state and delta updates in MRP
It is not possible to trust the playbackState field for playback
blindly. Some apps (like the stock Music app) reports "paused" even
when playing. To mitigate this, the playbackRate field is looked at and
if is is not 0, then we assume something is playing. But also fallback
to playbackState if missing.

Delta updates (updateContentItemMessage) are now merged into existing
items instead of completely overwriting them. This is how it's supposed
to be, otherwise existing metadata is lost.