Change AirPlay interface to Stream interface
Currently AirPlay is a bit of a bastard as it is possible to pair with
it but you cannot connect to a pure AirPlay service. The only thing
used in the AirPlay protocol is limited support for streaming media.
With this commit, AirPlay support is generalized in the interface to
just "Stream". This new interface is meant to provide functions to
stream media to a device, AirPlay just happens to be one way of doing
so. Next step is to make AirPlay less special.