mrp: Download protoc and verify generated code
The protobuf helper script will now download the protobuf compiler if
--download is specified. A version is hardcoded into the script so the
same version will always be used (a check is made for this as well).

By passing verify, the script will make sure that the generated code is
up-to-date. This check is buil into tox as well.

Examples:

  ./scripts/protobuf.py --download generate
  ./scripts/protobuf.py --download verify

If protoc has already been downloaded, it will not be downloaded agin.
Use --force to do it anyway.