scan: Wait for pending tasks
When a service is discovered, handling of the service is scheduled on
the event loop. This creates a race condition if a service is discovered
at the timeout border and the service will be missed. This is usually
not a problem but for tests, with stubbed sleep, this creates
instability. With this commit, scan will now wait for any pending
service after wait period has passed.