scripts: Add start of fake device script
This script will set up a fake device called FakeMRP at localhost. It
will be in idle state, but supports pairing and all commands that the
fake device implements. Most things are just logged so nothing really
happens.

A new instance is created for each connection to make sure encryption
works properly. Verified to work with the Home Assistant component.