mrp: Fix broken protobuf messages
I added an Unknown field a while back that caused the protobuf
generatation to fail due to name collision. This was unfortunate on one
end that message generation is not checked by tox, but also that
protobuf uses a "global" namespace for enums. To work around this I
moved the affected enums to individual messages, only for sake of scope.
This also made it easy to share types between message, which was an
issue before. The same thing should also be done for other messages
over time.