mrp: Potential fix for protocol hangs
The connection sometimes seems to hang after artwork has been received.
Reason is still unknown, but this change seems to make it work better.
It comes with the price that if artwork is large (file size), the
artwork might only be delivered as a transaction. Since transactions are
not implemented yet, artwork will fail in these cases.

Relates to #609.