Improve DNS message parsing for mDNS (#899)
* mdns: Start improving typing

The QTYPE_* constants are now members of the QueryType IntEnum, and the
old, untyped namedtuple are now typed NamedTuple.

* mdns: Import just 'typing' package

* mdns: Alphabetize stdlib imports

* mdns: Update pylint ignore directives for Exception

* mdns: Improve DNS-SD RFC conformance for properties

According to RFC 6763, the properties encoded in TXT records have ASCII
string keys, but the values are binary data. Additionally, the keys are
case-insensitive. A new class (CaseInsensitiveDict) was added for
storing properties. The only constraint is that the keys are strings, so
it's used for both the properties as decoded from the TXT record, and
and for the properties after the values have been stringified. To keep
it simpler, just the minimal methods have been overridden, and they're
simply lower-casing the keys before calling the base dict methods.

Using an actual case-insensitive dict from some other library might be a
better call later on.

* mdns: Improve type hints for tests and utils

Adding annotations as documentation in a few places, and fixing some
incorrect annotations as well.

* mdns: Improve name encoding RFC conformance

The actual maximum length of a label (aka DNS name component/level) is
63 octets, not 255. Also using IDNA for encoding the names to
(hopefully) handle internationalized domain names better.

* mdns: Rewrite DNS decoding to use BytesIO and improve RFC conformance

Parsing the DNS responses now uses a BytesIO instance instead of passing
around offsets and subscripted bytes objects. This was primarily done to
make the function signatures easier to reason about, but might also help
with performance (depending on how Python handles subscripting buffers).

The biggest improvement to standards conformance was by splitting
"domain-name" and "character-string" decoding. "domain-name"s are a
series of labels, each label up to 63 bytes, and there's also possible
name compression. The contained data also has restrictions as to what
can be present (basically ASCII characters). "character-string"s are
chunks of up to 255 bytes, with no restriction as to the contents. There
is no compression. The previous implementation conflated them, which
might've led to issues where a TXT label exceeded 63 bytes.

TXT records for DNS-SD parsing was also improved by adding support for
boolean keys (basically the key is just present, and has no "="
character).

There're also more type hints present, mainly on the updated functions.

* mdns: Add docstrings to CaseInsensitiveDict and QueryType

Mainly fixing linting issues.

* mdns: Add unpack_stream helper

It's struct.unpack, but for binary streams.

* mdns: Split DNS parsing and CaseInsensitiveDict into separate modules

* mdns: Add tests for dns module and fix bug with name decoding

* Loosen Service.properties type hints to Mappings

A future commit is going to be changing the parsed properties to be
MutableMappings, which are *not* Dicts. Keeping this in a separate
commit to make review easier.

* mdns: CaseInsensitiveDict is now a MutableMapping

While writing tests, some corner cases were revealed in the dict-based
implementation. By subclassing MutableMapping, the corner cases were
removed by eliminating the spots where dict modified its values
internally (ex: `update()` not calling `__setitem__()`).

* dns: Clarify docstring for unpack_stream()

* dns: Log a warning for name labels that are too long

Those labels are being truncated, but log a warning for them.

* dns: Use pack() methods on DnsHeader and DnsQuestion

Also modified DnsMessage.pack to use a `bytesarray` to avoid frequent
`bytes` creation (for every append, as `bytes` are immutable).

Co-authored-by: Pierre Ståhl <pierre.staahl@gmail.com>