gha: Generate html logs for logs in issues
When using the following syntax to add logs in an issue:

```log
log goes here
```

A workflow will be triggered that generates an html log, commits it to a
branch and posts a new comment in the issue with a link to the file.
It works with new issues, comments and edits to existing comments.