log2html: Fix broken log entry loading
It seems like onload does not work reliably with htmlpreview, so now a
polling loop looking for the element used for displaying log entries
is used instead. Not so pretty but works for now.

A workaround, storing the issue/comment body in a file instead of
passsing via environment variable has also been introduced. Storing too
large payloads in an environment variable seemed to throw an error.