Support for the Companion link protocol (#657)
* companion: Add stub for pairing

Relates to #655.

* companion: Add OPACK module

OPACK is a serialization format used by Companion link. This commit adds
basic support for packing and unpacking such data.

Relates to #655.

* opack: Change endianness to little

I was wrong about it being big endian, so changing to little.

Relates to #655.

* companion: Add support for pairing

Relates to #655.

* atvremote: Support companion credentials

Relates to #655.

* opack: Add support for additional types

UUUID4, endless lists and dicts as well as pointers are now supported.

Relates to #655

* companion: Initial support for Companion

This commit brings support for pairing and credential verificatation to
set up encryption for further communication. Only a basic API class is
provided, which isn't used yet.

Fixes #655

* companion: Update protocol documentation

Relates to #655

* atvproxy: Initial support for Companion

The proxy works per se, but the remote device performs an additional identity
validation which fails and closes the connection. So it is of very limited use
right now. Hopefully thks can improve over time.

Relates to #655

* support: Move chacha20 to support

Relates to #655

* support: Consolidate SRP for HAP implementations

SRP is 99% the same for MRP and Companion, so this change makes the code
generic enough to be shared by both protocols.

Relates to #655

* mrp: Lift changes from server_auth in Companion

I did a few changes in the Companion version of server_auth that I felt
I wanted to backport to MRP. So this commit fixes that and a few other
clean ups.

Relates to #655

* atvproxy: Tidy up MRP code

Relates to #655

* docs: Remove internal Companion documentation

Relates to #655

* cq: Fix minor code issues

Relates to #655