facade: Properly implement support for Features
The switch to use a facade implementation of the interface broke the
features interface (and was intentionally left to a separate commit).
With this commit, each protocol report what features (FeatureName) they
implement, which the feature relay implementation then uses to route
calls to the correct protocol. If multiple protocols support the same
feature, the protocol with highest priority will be used.

Relates to #1021