raop: Add initial support for RAOP (#1060)
* raop: Initial implementation of RAOP

Relates to #1059

* raop: Add shell support for RAOP

This commit makes it possible to scan for RAOP services and call the new
stream_file method in the Stream interface. There's no RAOP protocol
implementation yet, so it's not usable.

Relates to #1059

* raop: Add initial implementation of stream_file

This implementation is merely to demonstrate the functionality.

Relates to #1059

* support: Allow non-bytes arguments to log_binary

* raop: Add some additional debug logs

Relates to #1059

* raop: Send with additional packets when late

When being late sending packets, send some additional packets in an
attempt to catch up. A maximum of three packets will be sent each packet
interval to not overcompensate too much. A very naive approach but
seems to improve the situation a lot.

Fixes #1061

* raop: Send keep-alive message

Fixes #1062

* raop: Rewrite stream statistics

Refactor how statistics are calculated and provide better output for
easier troubleshooting. One small step towards improved statistics.

Relates to #1059

* raop: Use time.monotonic for time related things

Relates to #1059

* raop: Fixes for python 3.6

Relates to #1059

* raop: Add unit tests for packets

Relates to #1059

* docs: Update documentation for AirPlay and RAOP

I will add additional text to protocol documentation for RAOP and
AirPlay over time, but it's not a priority as there are lots of sources
online anyways.

Relates to #1059

* atvremote: Support manual RAOP config

Relates to #1059