tests: Support multiple addresses in MDNS
Normally, multiple addresses can be returned in a DNS response used for
MDNS, e.g. IPv4 and IPv6 addresses. Internally, pyatv only supports one
address (and that's still the case), but it has also been neglected in
tests. This is a refactoring step to support multiple addresses in a
single response as a prepration for filtering out unwanted addresses,
e.g. link-local.

Relates to #1090