conf: Add device info for AirPort Express (#1096)
* conf: Handle properties being convertde to lowcase

A side-effect of CaseInsensitiveDict was that the key-part of Zeroconf
properties were converted to lower-case. This passed by unnoticed, so
added a test for it. It's ok since it's very unlikely that two
properties with the same name, only differing in case, will ever exist.

Relates to #1090

* conf: Add detection of AirPort OS

Relates to #1090

* scan: Extract MAC and version for AirPort Express

This introduces "generic" properties in conf.AppleTV, extracted from
additional services not belonging to a certain protocol. The code is
currently very complex and needs a re-write, but this addition is
necessary to figure out the software version.

Fixes #1090