raop: Verify using AirPlay if credentials provided
RAOP and AirPlay (which in reality is the same protocol suite) uses the
same credentials to verify a connection. When playing on an Apple TV,
the RAOP connection must be verified (authenticated) or the stream won't
play. This change hi-jacks the verification process in AirPlay and
performs verification in case credentials are provided to RAOP. It will
do this regardless if the device supports it or not (as I don't know how
to verify if this process is needed) if credentials are provided.

Pairing needs to be done using AirPlay to obtain credentials. The same
credentials then needs to be provided to both AirPlay and RAOP.

Relates to #1078