raop: Fix timing when starting to stream
Current NTP time is stored in the context when setting up a stream and
if it takes a considerable amount of time to get the first audio sample,
then RTP time and real time will go out of sync. No audio. By resetting
and storing current NTP time just before starting to stream, we this
race condition does not exist and it will work.
Relates to #1161