raop: Sync sending based on absolute time
Prior to this commit, time between sending an audio packet was synced
per packet. From now on, an absolute time of current packet is
calculated and sleep time calculated according to where in time we
actually are. This is much more correct and will also adapt to system
load in a better way.

Relates to #1169