dmap: Add support for HSCP service
This commit adds support for the HSCP service used by the Music app (on
macOS). It's exactly the same as DMAP with Home Sharing, just a few
different zeroconf properties.

As of macOS 11.4 (Music 1.1.5.74), FairPlay authentication is required.
This change does not support that as algorithm has not been reverse
engineered yet.

Relates to #1172