airplay: Generate new credentials for Pair-Setup
It has generally been possible to provide credentials (e.g. seed and
identifier) via credentials in the service when pairing, thus not
generating new random credentials in those cases. This is no longer
possible for AirPlay and other protocols will follow the same pattern
in due time. It makes little sense to support this, as pairing should
generate new credentials and always doing that makes code simpler.

Relates to #1255