core: Allow setup to yield multiple instances
Under normal circumstances, one protocol implementation is used to
connect to one endpoint of that protocol. Now, when MRP is tunneled over
AirPlay, it will be the case that AirPlay sets up MRP. This change
allows re-use of the framework already in place to do that with very
hacks.

Relatest to #1255