conf: Move device_info implementations to protocol
This change defines a device_info method in each protocol used to
extract device information. The device_info method is called for each
service the protocol depends on and return only as much information as
it can possibly extract. All information is then merged at the end to
create the final device information.

Separate device_info methods are used for scanning and while connecting
to a device. The device_info used when connecting currently just calls
the one used by scanning but this allows additional device info to be
returned that has been obtained after connecting.

Relates to #1203