gha: Create GitHub release when tag is pushed
Today I manually create releases when I make a release, something that
should be automatic. This is an attempt to create automatic releases
when a tag is pushed. It is assumed a tag is always to be treated as a
release. As a test, releases will be marked as draft until I know that
it works.