if: Restructure service handling
This commit bundles a few changes:

* BaseService is now an abstract class and cannot be instansiated
* conf.ManualService has been added that implements BaseService
* All service specific implementations (e.g. MrpService) are deprecated
* A mutable version (MutableService) is used internally by pyatv

The intention here is to make service handling more generic and allow
internal parts of pyatv to add information to a service in steps,
without opening the mutable properties to the public interface. So far
this is not used in any way, but it will be used to support extra fields
for pairing and password requirements. Possibly other fields in the
future.

Relates to #1138