core: Use connect order based on protocol
Instead of connecting to protocols based on the order they were
discovered, use a pre-defined order (protocol name). It seems some Apple
TVs upgraded to tvOS 15 gets a lingering MRP service not possible to
connect to. If the MRP service is discovered prior to AirPlay,
connecting to such a device will fail. Maintaining this particular order
will serve as a work-around for that.

Relates to #1322