companion: Events and persistent connection
This commit contains a total re-write of the message dispatching in
Companion to support events (for OPACK frames). So far events are only
dispatched back to the API implementation (for media control
availability, i.e. which buttons are available) and not handled in any
way.

A persistent connection is required to support events, so from now on
a connection will be established when connecting (instead of just when
calling a function in the API). Sending system information is required
for this, otherwise the connection will be closed after a while.

Relates to #1192