core: Don't connect to dummy services
When setting up MRP-over-AirPlay, a dummy service is inserted into the
configuration for MRP. This is needed when setting up MRP. It however
created a problem in the particular case when a remote control is set up
but fails to connect. This happens for Apple TVs as transient pairing
is assumed (as the device announce support for that), which then fails.
This triggers a new connect attempt to "regular" MRP with the dummy
service, which obviously also fails and triggers a hard error.

Relates to #1357