airplay: Quiet log of remote control setup failure
Whenever setting up a remote control connection (MRP-over-AirPlay) that
fails, a warning log is triggered. It's annoying and should not be
there, so this commits changes it to a debug log. It will however only
do so for error code 470 and logs exception for everything else. This
should give better error messages for actual errors as well as being
less noisy in normal cases.