env: Re-work test environment
This commit bringa a _lot_ of changes to how pyatv is tested and
verified by GitHub Actions. First and foremost: tox has been replaced
by a new tool called chickn. It's a simple automation tool to run
commands in a similar fashin to tox, but way more lightweight and fitted
for what pyatv is doing. This gives me more flexibility but has also
decreased the general test time a bit. See documentation for more
details and how to use chickn. Do not that tox has been completely
ripped out.

The switch to chickn has made it easier to cache the virtual environment
in which pyatv is tested, resulted in further reduced test times. For
some reason I could not get this stable with tox, most time the caches
wouldn't work.

An additional caching technique has been added to not re-run jobs that
have already passed when re-running the "tests" workflow. This also
greatly reduce test times in case of a flaky test or problems with the
test environment.

Tests are no longer run for python 3.6 on Ubuntu as they are really,
really unstable. That combination is still supported by pyatv, just not
tested as it fails about half of the time. The idea is that tests run
on Windows and macOS catch most issues anyways. At some point I will
drop support for python 3.6, so it's not worth the trouble.

Regression is tested as a separate job now.

Requirement files have been moved to the requirements directory and will
always contain the latest versions of everything (it is what pyatv is
tested with). The base_versions.txt file still remains in the root and
lists the minimum required versions of what pyatv supports. This is not
bumped by dependabot.

Relates to #1282