core: Add base support for internal state updates
The idea is to allow protocols to dispatch and subscribe to state
changes internally, e.g. when the playing state changes someone else can
subscribe to that and perform an action when it happens. In the long run
this is a preparation for a more flexible listener interface (where
listener classes won't be needed). To some extent there will be overlaps
with StateProducer, but I intend to fix that over time.

Relates to #1500