facade: Make Power interface respect play status
This change subscribes to play state updates in the facade power
implementation and will override power state (to "On") if something is
playing regardless of what the underlying power implementation thinks.
None of the protocols dispatch the play state yet, so this change does
nothing at the moment.

Relates to #1500