core: Prevent duplicates from state producer
Duplicate push updates shall never be sent, i.e. the same state shall
not be published several times in a row. This was prevented for regular
updates via StateProducer but not via StateDispatcher, which resulted in
power updates being published more than once (since it depends on play
state updates now).

Relates to #1500