Add support for URL param to LaunchApp
This change adds support for passing the optional, undocumented `_urlS`
parameter to `_launchApp`. Passing this parameter enables deep linking
into at least some apps installed on the AppleTV. I discovered this
while searching for a way to launch content directly from a deep link
into the AppleTV+ app.

The first clue that this param exists was that passing a valid URL in
 place of a bundleID returns the error message "No valid bundleID or
URL to LaunchApp". Searching for strings in the rootfs of the tvos 15.6
image for AppleTV5,3 (build 19M65) suggested that the `rapportd` daemon
is responsible for handling companion protocol requests and returning
this error message.

A hidden `_urlS` string key is present in the `rapportd` daemon located near
`_bundleID`:

```
$ strings rapportd | grep -C 5 bundleID
_launchApp
v32@?0@"NSDictionary"8@"NSDictionary"16@?<v@?@"NSDictionary"@"NSDictionary"@"NSError">24
_sessionStart
_sessionStop
v40@?0@"RPConnection"8@"NSDictionary"16@"NSDictionary"24@?<v@?@"NSDictionary"@"NSDictionary"@"NSError">32
_bundleID
RPLaunchApp
_urlS
RPOpenURL
-[RPCompanionLinkDaemon _miscHandleLaunchAppRequest:responseHandler:]_block_invoke
```

Disassembly of the object code reveals the following logic in the
`-[RPCompanionLinkDaemon _miscHandleLaunchAppRequest:responseHandler:]` method:

```
(pseudocode)
if (_bundleID):
    launch_app(_bundleID)
elif (_urlS):
    open_url(_urlS)
else:
    error("No valid bundleID or URL to LaunchApp")
```

If a bundleID is passed to `_launchApp`, it takes precendce. The implementation
provided here does not attempt to enforce this logic on the client. Instead,
params are passed directly through to the server so that there is only one
source of truth for the behavior. The current server behavior is represented
in the updated companion protocol tests.