airplay: Fix play_url with newer tvOS versions
The play_url method has been broken for a very long time as I believed
Apple wasn't maintaining the underlying functionality anymore. They do,
but an AirPlay session must be established for it to work now. So I
have added that and verified with an Apple TV 3 and 4K running
tvOS 14.5.

Relatest to #2064