airplay: Add setting for MRP tunnel
It is now possible to force or disable set up of MRP over AirPlay tunnel
over Airplay (or the default "auto" mode). If set up of remote control
fails, that will now render an exception (just like with other protocols
on an error) instead of just a log message.

Relates to #2629