companion: Use static id for _i in _systemInfo
This change generates a random "rp identifier" that is included in _systemInfo
sent by the Companion protocol. It's not really known what this identifier is,
but it comes from rapportd/Rapport.framework and seems to be unique per device.
So we generate a unique one per device and store it with settings.

Adjustments have been made so that if this identifier is missing in settings,
a new identifier will be generated and saved automatically.

Relates to #2656